deployment/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.50.0
toolhouse>=1.4.0
python-dotenv>=1.0.1
pydantic>=2.5.0
httpx>=0.27.0 